Update Jobbr nuget packges and fix integration
