+++ source/Sandbox.JobServer/Sandbox.JobServer.csproj
@@ -14,14 +14,17 @@
     <DebugType>pdbonly</DebugType>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="Jobbr.Server" Version="3.0.0-rc.2" />
+    <PackageReference Include="Jobbr.Server" Version="3.0.0-rc.4" />
     <PackageReference Include="Jobbr.Execution.Forked" Version="3.0.0-rc.1" />
     <PackageReference Include="Jobbr.Storage.MsSql" Version="3.0.0-rc.3" />
     <PackageReference Include="ServiceStack.OrmLite.SqlServer" Version="6.0.0" />
 
     <PackageReference Include="Jobbr.ComponentModel.Management" Version="3.0.0-rc.1" />
-    <PackageReference Include="Jobbr.ComponentModel.Registration" Version="3.0.0-rc.1" />
+    <PackageReference Include="Jobbr.ComponentModel.Registration" Version="3.0.0-rc.2" />
+    <PackageReference Include="Jobbr.ComponentModel.JobStorage" Version="3.0.0-rc.1" />
     <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="7.0.0" />
+    <PackageReference Include="Jobbr.ComponentModel.ArtefactStorage" Version="3.0.0-rc.1" />
+    <PackageReference Include="Jobbr.ComponentModel.Execution" Version="3.0.0-rc.1" />
     <PackageReference Include="ncrontab" Version="3.3.1" />
   </ItemGroup>
   <ItemGroup>
